requirements.txt
pandas
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
ipykernel
openpyxl
langchain-google-genai
